requirements.txt
python-dotenv
requests